query/requirements.txt
fastapi
pydantic
langchain
langchain-community
faiss-cpu
pandas
numpy
sentence_transformers
uvicorn
pymorphy2